+++ packages/perseus-core/src/index.ts
@@ -27,6 +27,7 @@ export {
 } from "./utils/widget-id-utils";
 export {default as deepClone} from "./utils/deep-clone";
 export * as GrapherUtil from "./utils/grapher-util";
+export {generateTestPerseusItem} from "./utils/test-utils";
 
 export {
     parsePerseusItem,