+++ packages/perseus/src/__tests__/test-items/mock-widget-2-item.ts
@@ -23,9 +23,4 @@ export default {
     },
     answerArea: null,
     hints: [] as ReadonlyArray<any>,
-    itemDataVersion: {
-        major: 1,
-        minor: 0,
-    },
-    answer: null,
 } satisfies PerseusItem;