+++ packages/perseus/src/util/test-utils.ts
@@ -6,7 +6,6 @@ import type {
     ExpressionWidget,
     InteractiveGraphWidget,
     NumericInputWidget,
-    PerseusItem,
     PerseusRenderer,
     PerseusWidgetTypes,
     RadioWidget,
@@ -14,31 +13,6 @@ import type {
     PerseusScore,
     UserInputMap,
 } from "@khanacademy/perseus-core";
-
-export const genericPerseusItemData: PerseusItem = {
-    question: {
-        content: "",
-        images: {},
-        widgets: {},
-    },
-    answerArea: {
-        calculator: false,
-        chi2Table: false,
-        periodicTable: false,
-        tTable: false,
-        zTable: false,
-        financialCalculatorMonthlyPayment: false,
-        financialCalculatorTotalAmount: false,
-        financialCalculatorTimeToPayOff: false,
-        periodicTableWithKey: false,
-    },
-    itemDataVersion: {
-        major: 0,
-        minor: 1,
-    },
-    hints: [],
-    answer: null,
-} as const;
 
 /**
  * Thin wrapper around scorePerseusItem for internal testing
@@ -48,21 +22,6 @@ export function scorePerseusItemTesting(
     userInputMap: UserInputMap,
 ): PerseusScore {
     return scorePerseusItem(perseusRenderData, userInputMap, "en");
-}
-
-/**
- * Generate a Perseus item object for testing purposes.
- *
- * In order to better type Perseus objects used in testing, this function
- * uses a basic Perseus object and updates it with custom values as needed.
- *
- * @param {Partial<PerseusItem>} customFields
- * @returns {PerseusItem}
- */
-export function generateTestPerseusItem(
-    customFields: Partial<PerseusItem> = {},
-): PerseusItem {
-    return {...genericPerseusItemData, ...customFields};
 }
 
 /**