+++ packages/pure-markdown/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @khanacademy/pure-markdown
 
+## 2.0.2
+
+### Patch Changes
+
+-   Updated dependencies []:
+    -   @khanacademy/perseus-utils@2.0.1
+    -   @khanacademy/simple-markdown@2.0.2
+
 ## 2.0.1
 
 ### Patch Changes